reorganized the testing data and resources
