+++ cmstk/eam_test.py
@@ -1,13 +1,12 @@
 from cmstk.eam import SetflFile
+from cmstk.testing_resources import data_directory
 import os
 import numpy as np
 
 def test_setfl_file():
     """Test the initialization of an eam.SetflFile class."""
-    path = os.path.abspath(__file__)
-    path = os.path.dirname(path)
-    path = os.path.dirname(path)
-    path = os.path.join(path, "data", "potentials", 
+    path = os.path.join(data_directory(),
+                        "potentials",
                         "Bonny-Fe-Ni-Cr-2011.eam.alloy")
     setfl = SetflFile(path)
     setfl.read()
