+++ cmstk/units/energy_test.py
@@ -1,5 +1,5 @@
 from cmstk.units.energy import EnergyUnit, ElectronVolt, Joule
-from cmstk.units.testing_resources import within_one_percent
+from cmstk.testing_resources import within_one_percent
 
 # ElectronVolt
