+++ cmstk/units/vector_test.py
@@ -1,7 +1,7 @@
 from cmstk.units.vector import Vector, Vector2D, Vector3D
 from cmstk.units.distance import DistanceUnit, Angstrom, Nanometer, Meter, Picometer
 from cmstk.units.angle import AngleUnit, Radian, Degree
-from cmstk.units.testing_resources import within_one_percent
+from cmstk.testing_resources import within_one_percent
 import pytest
 import numpy as np
