+++ data/atat/bestsqs.out
