+++ data/atat/rndstr.in
